requirements.txt
# Core dependencies
pandas==2.0.3
requests==2.31.0
urllib3==2.2.1
python-dotenv==1.0.1

# SQLAlchemy for DB connection
sqlalchemy>=1.4,<2.0
psycopg2-binary==2.9.9

# Data validation
pandera==0.18.0

# Testing
pytest==7.4.4

# Data visualization
matplotlib==3.7.1